# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Bible API
  version: 1.0.0
  description: A comprehensive RESTful Bible API with AI integration
paths:
...
  /api/v1/bible/books/:
    get:
      operationId: bible_books_list
      description: |-
        Mixin to automatically add Vary: Accept-Language header to all responses.

        This ensures proper HTTP caching behavior for internationalized endpoints
        that may return different content based on Accept-Language header.

        Usage:
            class MyView(LanguageSensitiveMixin, APIView):
                pass

        The mixin works by overriding finalize_response to add the Vary header
        to all response types (2xx, 4xx, 5xx).
      parameters:
      - in: query
        name: is_deuterocanonical
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: testament
        schema:
          type: integer
      tags:
      - bible
      security:
      - ApiKeyAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBookList'
          description: ''
...
components:
  schemas:
...
    Book:
      type: object
      description: Serializer for canonical books with backward compatibility.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          description: Get localized name using the existing utility function with
            fallback logic.
          readOnly: true
        abbreviation:
          type: string
          description: Get localized abbreviation using the existing utility function
            with fallback logic.
          readOnly: true
        order:
          type: integer
        testament:
          type: string
        chapter_count:
          type: integer
          maximum: 2147483647
          minimum: 0
      required:
      - abbreviation
      - chapter_count
      - id
      - name
      - order
      - testament
...
    PaginatedBookList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Book'
...